+++ src/main/scala/com/databricks/labs/gbx/rasterx/expressions/accessors/RST_Min.scala
@@ -54,21 +54,12 @@ object RST_Min extends WithExpressionInfo {
     def execute(ds: Dataset): Array[java.lang.Double] = {
         (1 to ds.GetRasterCount()).map { bandIndex =>
             val band = ds.GetRasterBand(bandIndex)
-            if (band == null) null
-            else {
-                val md = band.AsMDArray()
-                val stats = md.GetStatistics()
-                val res: java.lang.Double =
-                    if (stats == null || stats.getValid_count == 0) null
-                    else {
-                        val (min, _) = BandAccessors.getMinMax(band)
-                        min
-                    }
-                if (stats != null) stats.delete()
-                md.delete()
-                band.delete()
-                res
-            }
+            val res: java.lang.Double =
+                if (band == null) null
+                else if (BandAccessors.isEmpty(band)) null
+                else java.lang.Double.valueOf(BandAccessors.getMinMax(band)._1)
+            if (band != null) band.delete()
+            res
         }.toArray
     }
